Fuss naming and structure, use DeserializeOwned trait for data handling
